FEAT: добавлена база данных feat: добавлено взаимодействие с БД delete: удалены имплементации интерфейсов репозиториев feat: обновлены сервисы, контроллеры и модели

+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -11,11 +11,13 @@ import javax.validation.constraints.NotNull;
 @AllArgsConstructor
 @Builder
 public class ItemDto {
-    private Integer id;
+    private Long id;
     @NotBlank
     private String name;
     @NotBlank
     private String description;
     @NotNull
     private Boolean available;
+
+    //TODO добавить комменты в dto
 }
